get_bay_yelp_data/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
requests
datetime